Enable torrent creation via the api
Closes #5614.

Implement a new api controller to handle generation of .torrent files.
The jobs to create torrents on API side are bound to a session. It's
not possible to access the jobs of a different session; and pending
creations are aborted when destroying the session. It's a lazy abort:
objects will remain queued to the QThreadPool, but marked as
"interruption requested", so will be skipped rapidly during execution.

Slightly adapt the TorrentCreator for the new needs: allow to return
the content of the generated .torrent file without requiring to write
it to disk at all. Also, communicate the actual pieceSize used for
generation (if it's set to 0, libtorrent choses it automatically).