Enable torrent creation via the api
Implement a new api controller to handle generation of .torrent files.
Add a new TorrentCreationManager singleton which keeps tracks of
torrent creation tasks. Use a separate ThreadPool in the Manager.
It acts as a queueing mechanism for creation tasks to avoid too many
tasks running in parallel.

Slightly adapt the TorrentCreator for the new needs: allow to return
the content of the generated .torrent file without requiring to write
it to disk at all. If no savePath is passed, the created torrent file
will be kept in memory. Also, communicate the actual pieceSize used for
generation (if it's set to 0, libtorrent selects it automatically).

Closes #5614.