Implement SSL peer support
This feature is interesting when multiple qBittorent instances
are managed by the same administrator who uses it to transfer private
data across a non-secure network.

Implement the two missing functionalities:
- allow to configure and listen on an SSL port
- allow to set the SSL certificate on a torrent via the API

More details about these two functionalities:

A separate port has to be allocated for incoming SSL connections from
peers. Libtorrent already supports this. It's enough to add the
suffix 's' when configuring libtorrent's listen_interfaces. Implement
all the gui and core changes to allow setting an SSL port on top of the
currently existing normal port.

Configuring an SSL port is not enough to allow SSL connection. x509
certificate, private_key and diffie-hellman parameters have to be
configured for each torrent. This is achieved by calling libtorrent's
handle->set_ssl_certificate. The certificates are only kept in-memory,
so they have to be explicitly re-added after each restart.

Because of the constraints listed previously, some automation on top
of qBittorent's API is strongly desired to make full use of this new
feature. An alternative way to set those certificates is to put them
in a directory. The certificates from that path will be automatically
added to affected torrents when torrent_need_cert_alert is raised.
Certificates have to be stored in that path with pre-defined file names.